Lesson 69 (commit 8) change fields of class and files for fix project

+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/Repository/ProductsRepository.java
@@ -7,10 +7,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ProductsRepository  extends PagingAndSortingRepository<Products,Integer> {
+public interface ProductsRepository extends PagingAndSortingRepository<Products, Integer> {
 
     @Query("SELECT p FROM Products p WHERE CONCAT(p.name, ' ', p.description, ' ', p.price) LIKE %?1%")
     Page<Products> search(String keyword, Pageable pageable);
 
-    Page<Products> findProductsByPriceGreaterThanAndPriceIsLessThan(Integer less,Integer more, Pageable pageable);
+    Page<Products> findProductsByPriceGreaterThanAndPriceIsLessThan(Float less, Float more, Pageable pageable);
+
+
 }
